move: Home components to components package
