+++ app/src/main/kotlin/com/yourssu/soomsil/usaint/screen/home/components/ChapelCardItem.kt
@@ -1,10 +1,10 @@
-package com.yourssu.soomsil.usaint.screen.home
+package com.yourssu.soomsil.usaint.screen.home.components
 
-
-/*@Composable
-private fun GraduationCard(
+/*
+@Composable
+private fun ChapelCardItem(
     isLoggedIn: Boolean,
-    onGraduationCardClick: () -> Unit,
+    onChapelCardClick: () -> Unit,
 ) {
     Surface(
         rounding = 8.dp,
@@ -16,7 +16,7 @@ private fun GraduationCard(
                 .padding(bottom = 10.dp),
         ) {
             YdsText(
-                text = stringResource(id = R.string.saint_graduation),
+                text = stringResource(id = R.string.saint_chapel),
                 modifier = Modifier
                     .padding(
                         top = 20.dp,
@@ -30,9 +30,16 @@ private fun GraduationCard(
                 ),
             )
             ActionTitle(
-                title = stringResource(id = R.string.saint_graduation_title),
-                subTitle = stringResource(id = R.string.saint_graduation_subtitle),
-                onClick = onGraduationCardClick,
+                title = stringResource(id = R.string.saint_chapel_seat_title),
+                subTitle = stringResource(id = R.string.saint_chapel_seat_subtitle),
+                onClick = onChapelCardClick,
+                enable = false,
+                enable = isLoggedIn,
+            )
+            ActionTitle(
+                title = stringResource(id = R.string.saint_chapel_attendance_title),
+                subTitle = stringResource(id = R.string.saint_chapel_attendance_subtitle),
+                onClick = onChapelCardClick,
                 enable = false,
                 enable = isLoggedIn,
             )
